+++ docs/app/models/employee.rb
@@ -1,11 +1,6 @@
-  # frozen_string_literal: true
+class Employee < ApplicationRecord
+  include ActiveSupport::NumberHelper
+  has_and_belongs_to_many :locations
 
-  class Employee
-    attr_reader :first_name, :last_name, :email, :phone
-    def initialize(first_name:, last_name:, email:, phone:)
-      @first_name = first_name
-      @last_name = last_name
-      @email = email
-      @phone = phone
-    end
-  end
+  def formatted_phone = number_to_phone(phone, pattern: /(\d{3})(\d{4})(\d{4})$/)
+end