+++ lib/rui/table.rb
@@ -7,28 +7,21 @@ class RUI::Table < RUI::Base
   def view_template(&)
     vanish(&)
 
-    table(class: "bg-white w-full table-auto border-collapse") do
+    table(class: "bg-white w-full table-auto border-collapse rounded") do
       thead(class: "bg-blue-50") do
         @columns.each do |column|
-          th(class: "border border-zinc-300 py-2 px-2 #{align(column[:align])}") do
+          th(class: classes(column[:attrs][:class])) do
             column[:header]
           end
         end
-        th(class: "border border-zinc-300 px-2 text-center") { "" }
       end
 
       tbody do
         @rows.each do |row|
-          tr(class: "hover:bg-zinc-50") do
+          tr(class: "odd:bg-white even:bg-zinc-100 hover:bg-zinc-200") do
             @columns.each do |column|
-              td(class: "border border-zinc-300 py-1 px-2 #{align(column[:align])}") do
+              td(class: classes(column[:attrs][:class])) do
                 column[:content].call(row)
-              end
-            end
-            td(class: "border border-zinc-300 py-1 px-2 text-center") do
-              div(class: "flex flex-row justify-center gap-3") do
-                render RUI::Buttons::Ghost.new(icon: "edit")
-                render RUI::Buttons::Ghost.new(icon: "archive")
               end
             end
           end
@@ -37,18 +30,15 @@ class RUI::Table < RUI::Base
     end
   end
 
-  def column(header, align = :left, &content)
-    @columns << { header:, align:, content: }
+  def column(header = "", **attrs, &content)
+    @columns << { header:, attrs:, content: }
   end
 
-  def align(align)
-    case align
-    when :right
-      "text-right"
-    when :center
-      "text-center"
-    else
-      "text-left"
-    end
+  private
+
+  def base_classes = "border border-zinc-300 py-1 px-2 text-left"
+
+  def classes(custom_classes)
+    RUI::TailwindMerger.instance.merge(base_classes, custom_classes)
   end
 end