+++ lib/rui/tailwind_merger.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal
+
+require "tailwind_merge"
+
+class RUI::TailwindMerger
+  include Singleton
+  include TailwindMerge
+
+  def initialize
+    @merger = TailwindMerge::Merger.new
+  end
+
+  def merge(base_classes, given_classes)
+    return base_classes if given_classes.nil?
+    @merger.merge([ base_classes, given_classes ])
+  end
+end